+++ popkorn/src/main/kotlin/cc/popkorn/mapping/Mapping.kt
@@ -3,7 +3,7 @@ package cc.popkorn.mapping
 import kotlin.reflect.KClass
 
 /**
- * Interface that creates gets the cc.popkorn.mapping of a certain class
+ * Interface that maps any class to any object
  *
  * @author Pau Corbella
  * @since 1.1.0
@@ -11,9 +11,5 @@ import kotlin.reflect.KClass
 interface Mapping {
 
      fun find(original:KClass<out Any>) : Any?
-
-     fun isPresent(original:KClass<*>) : Boolean{
-          return find(original) != null
-     }
 
 }