+++ CHANGELOG.md
@@ -3,7 +3,8 @@ Change Log
 
 Version 1.3.0 *(XXXXXX)*
 -----------------------------
-* New: Runtime injectable can now be an interface
+* New: Custom `Injector` can use custom pools
+* Fix: Runtime injectable can now be an interface
 * New: Can get the current environment in any injectable constructor / method by defining `Environment` parameter
 * Fix: Multiple environments in a constructor/method result in an invalid compilation file